Allow choosing the solution from a dropdown menu of existing objects.
If an object is mistakenly flagged with the wrong value, the game
will be unable to guess it. This change allows the solution to be
picked from a dropdown instead of having to type the exact same name
into a text box (which obviously also still works).

In addition, this is a convenient way to update properties for objects
preceding said properties: reject the (correct) guess, flag all other
properties, then finally choose the object from the dropdown.